Improved Commit system and user agent management
