+++ Exception/UserAgentNotFoundException.php
@@ -1,0 +1,32 @@
+<?php
+
+/*
+ * Copyright (c) 2015 Lp digital system
+ *
+ * This file is part of Github Parser library.
+ *
+ * Github Parser is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * Github Parser is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with Github Parser. If not, see <http://www.gnu.org/licenses/>.
+ */
+
+namespace Lpdigital\Github\Exception;
+
+use Exception;
+
+class UserAgentNotFoundException extends Exception
+{
+    public function __construct()
+    {
+        parent::__construct('You need to set user agent to execute this query.', 0, null);
+    }
+}
